pyproject.toml
[project]
name = "fast-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "python-dotenv>=1.1.1",
]
